feat : 로그인 시 헤더 프로필 정보 불러오기
